fix(shared): allow updating only one file from map version

+++ apps/backend/src/app/modules/maps/maps.service.ts
@@ -539,6 +539,9 @@ export class MapsService {
 
     const bspFile = await this.getBspFromTemp(userID);
 
+    if (bspFile instanceof Error)
+      throw new BadRequestException('Missing BSP file');
+
     await this.checkMapCompression(bspFile);
 
     this.checkMapFiles(bspFile, vmfFiles);
@@ -652,14 +655,21 @@ export class MapsService {
       throw new ForbiddenException('Map does not allow editing');
     }
 
+    let bspHash: string;
+
     const bspFile = await this.getBspFromTemp(userID);
 
-    await this.checkMapCompression(bspFile);
+    if (bspFile instanceof Error) {
+      bspHash = map.currentVersion.bspHash;
+    } else {
+      await this.checkMapCompression(bspFile);
 
-    this.checkMapFiles(bspFile, vmfFiles);
+      this.checkMapFiles(bspFile, vmfFiles);
+
+      bspHash = FileStoreService.getHashForBuffer(bspFile.buffer);
+    }
 
     const hasVmf = vmfFiles?.length > 0;
-    const bspHash = FileStoreService.getHashForBuffer(bspFile.buffer);
 
     let zonesStr: string;
     let zones: MapZones;
@@ -723,7 +733,11 @@ export class MapsService {
               ? await this.zipVmfFiles(map.name, newVersionNum, vmfFiles)
               : undefined;
 
-            await this.uploadMapVersionFiles(newVersion.id, bspFile, zippedVmf);
+            await this.uploadMapVersionFiles(
+              newVersion.id,
+              bspFile instanceof Error ? undefined : bspFile,
+              zippedVmf
+            );
           })()
         );
 
@@ -998,21 +1012,23 @@ export class MapsService {
 
   private async uploadMapVersionFiles(
     uuid: string,
-    bspFile: File,
+    bspFile?: File,
     vmfZip?: Buffer
   ) {
     const storeFns: Promise<FileStoreFile | boolean>[] = [];
 
-    if (bspFile.path) {
-      storeFns.push(
-        this.fileStoreService
-          .copyFile(bspFile.path, bspPath(uuid))
-          .then(() => this.fileStoreService.deleteFile(bspFile.path))
-      );
-    } else {
-      storeFns.push(
-        this.fileStoreService.storeFile(bspFile.buffer, bspPath(uuid))
-      );
+    if (bspFile) {
+      if (bspFile.path) {
+        storeFns.push(
+          this.fileStoreService
+            .copyFile(bspFile.path, bspPath(uuid))
+            .then(() => this.fileStoreService.deleteFile(bspFile.path))
+        );
+      } else {
+        storeFns.push(
+          this.fileStoreService.storeFile(bspFile.buffer, bspPath(uuid))
+        );
+      }
     }
 
     if (vmfZip)
@@ -1891,13 +1907,13 @@ export class MapsService {
       throw new BadRequestException('BSP is not compressed');
   }
 
-  async getBspFromTemp(userID: number): Promise<File> {
+  async getBspFromTemp(userID: number): Promise<File | Error> {
     const userIDObjects = await this.fileStoreService.listFileKeys(
       `upload_tmp/${userID}`
     );
 
     if (userIDObjects.length === 0) {
-      throw new BadRequestException('Missing BSP file');
+      return new Error('Missing BSP file');
     }
 
     const bspData = await this.fileStoreService.getFile(userIDObjects[0]);
